Fix Group engine
