+++ engines/group/.rspec
@@ -1,2 +1,1 @@
 --color
---require spec_helper
