+++ engines/group/lib/group/engine.rb
@@ -16,7 +16,6 @@ module Group
     initializer "model_core.factories", after: "factory_girl.set_factory_paths" do
       if defined?(FactoryGirl)
         FactoryGirl.definition_file_paths << File.expand_path('../../../spec/factories', __FILE__)
-        FactoryGirl.reload
       end
     end
   end
