+++ engines/group/spec/dummy/config/database.yml
@@ -1,16 +1,17 @@
-# SQLite version 3.x
-#   gem install sqlite3
-#
-#   Ensure the SQLite 3 gem is defined in your Gemfile
-#   gem 'sqlite3'
-#
 default: &default
+  username: <%= ENV['DB_USERNAME'] || 'postgres' %>
+  password: <%= ENV['DB_PASSWORD'] %>
   adapter: postgresql
   pool: 5
+  encoding: unicode
 
 # Warning: The database defined as "test" will be erased and
 # re-generated from your development database when you run "rake".
 # Do not set this db to the same as development or production.
 test:
   <<: *default
-  database: katuma_group_test
+  database: katuma_test
+
+development:
+  <<: *default
+  database: katuma_development
