+++ engines/group/spec/models/group/user_spec.rb
@@ -1,0 +1,8 @@
+require 'rails_helper'
+
+describe Group::User do
+  describe 'Associations' do
+    it { should have_many(:memberships).dependent(:destroy) }
+    it { should have_many(:groups).through(:memberships) }
+  end
+end
